pyproject.toml
[build-system]
requires = ["hatchling>=1.28.0"]
build-backend = "hatchling.build"

[project]
name = "obsidian-vault-rag"
version = "0.1.0"
description = "Obsidian RAG CLI - Ingest and Chat with your notes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lancedb==0.26.1",
    "pyyaml==6.0.3",
    "sentence-transformers==5.2.0",
    "langchain-text-splitters==1.1.0",
    "langchain-docling==2.0.0",
    "tqdm==4.67.1",
    "pydantic==2.12.5",
    "mcp==1.25.0",
    "docling==2.67.0",
    "typer>=0.15.0",
    "rich==13.7.0",
    "einops",
]

[project.scripts]
obsidian = "obsidian.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian"]

[tool.ruff]
line-length = 120
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "PLE",    # pylint errors
    "PLR",    # pylint refactor
    "PLW",    # pylint warnings
    "T",      # flake8-print
]
ignore = [
    "PLW0603",  # global-statement (needed for singleton pattern)
    "E501",     # line-too-long (we use 120 chars)
]

[tool.ruff.lint.per-file-ignores]
"src/obsidian/cli.py" = ["T201", "T203"]  # Allow print statements in CLI

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
suppress-dummy-args = false

[tool.ruff.lint.isort]
known-first-party = ["obsidian"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 120

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "auto"

[project.optional-dependencies]
test = [
    "pytest>=9.0.2",
]
dev = [
    "ruff>=0.14.11",
]
